Bank Marketing Data-Analysis Project
Marketing Introduction:-
The process by which companies create value for customers and build strong customer relationships in order to capture value from customers in return.
Marketing campaigns are characterized by focusing on the customer needs and their overall satisfaction. Nevertheless, there are different variables that determine whether a marketing campaign will be successful or not. There are certain variables that we need to take into consideration when making a marketing campaign.
Machine Learning Steps Involved:-
1.	XGBoost

Steps Involved:-
1.	Import data from dataset and perform initial high-level analysis: 
2.	Data Exploration (look at the number of rows, look at the missing values, and look at dataset columns and their values respective to the campaign outcome).
3.	Data Cleaning (Cleaning the data: remove irrelevant columns)
4.	Dealing with missing and incorrect values, 
5.	Turning categorical columns into dummy variables.
6.	Use machine learning techniques to predict the marketing campaign outcome and to find out factors, which affect the success of the campaign. 


Dataset Information:-
The data is related with direct marketing campaigns of a Portuguese banking institution. The marketing campaigns were based on phone calls. Often, more than one contact to the same client was required, in order to access if the product (bank term deposit) would be ('yes') or not ('no') subscribed.
The classification goal is to predict if the client will subscribe (yes/no) a term deposit (variable y).